+++ processes/mit.py
@@ -1,0 +1,105 @@
+import json
+import os
+from requests.exceptions import HTTPError, ConnectionError
+
+from .core import CoreProcess
+from mappings.core import MappingError
+from mappings.mit import MITMapping
+from managers import WebpubManifest
+from logger import createLog
+
+logger = createLog(__name__)
+
+class MITProcess(CoreProcess):
+
+    def __init__(self, *args):
+        super(MITProcess, self).__init__(*args[:4])
+
+        self.ingestOffset = int(args[5] or 0)
+        self.ingestLimit = (int(args[4]) + self.ingestOffset) if args[4] else 5000
+        self.fullImport = self.process == 'complete' 
+
+        # Connect to database
+        self.generateEngine()
+        self.createSession()
+
+        # S3 Configuration
+        self.s3Bucket = os.environ['FILE_BUCKET']
+        self.createS3Client()
+
+    def runProcess(self):
+        with open('mitMetaData.json') as f:
+            mitData = json.load(f)
+
+        for metaDict in mitData:
+            self.processMITRecord(metaDict)
+
+        self.saveRecords()
+        self.commitChanges()
+
+    def processMITRecord(self, record):
+        try:
+            mitRec = MITMapping(record)
+            mitRec.applyMapping()
+            self.storePDFManifest(mitRec.record)
+            self.addDCDWToUpdateList(mitRec)
+            
+        except (MappingError, HTTPError, ConnectionError, IndexError, TypeError) as e:
+            logger.exception(e)
+            logger.warn(MITError('Unable to process MIT record'))
+            
+
+    def storePDFManifest(self, record):
+        for link in record.has_part:
+            itemNo, uri, source, mediaType, flags = link.split('|')
+
+            if mediaType == 'application/pdf':
+                recordID = record.identifiers[0].split('|')[0]
+
+                manifestPath = 'manifests/{}/{}.json'.format(source, recordID)
+                manifestURI = 'https://{}.s3.amazonaws.com/{}'.format(
+                    self.s3Bucket, manifestPath
+                )
+
+                manifestJSON = self.generateManifest(record, uri, manifestURI)
+
+                self.createManifestInS3(manifestPath, manifestJSON)
+
+                linkString = '|'.join([
+                    itemNo,
+                    manifestURI,
+                    source,
+                    'application/webpub+json',
+                    '{"catalog": false, "download": false, "reader": true, "embed": false}'
+                ])
+
+                record.has_part.insert(0, linkString)
+                break
+
+    def createManifestInS3(self, manifestPath, manifestJSON):
+        self.putObjectInBucket(
+            manifestJSON.encode('utf-8'), manifestPath, self.s3Bucket
+        )
+
+    @staticmethod
+    def generateManifest(record, sourceURI, manifestURI):
+        manifest = WebpubManifest(sourceURI, 'application/pdf')
+
+        manifest.addMetadata(
+            record,
+            conformsTo=os.environ['WEBPUB_PDF_PROFILE']
+        )
+        
+        manifest.addChapter(sourceURI, record.title)
+
+        manifest.links.append({
+            'rel': 'self',
+            'href': manifestURI,
+            'type': 'application/webpub+json'
+        })
+
+        return manifest.toJson()
+
+
+class MITError(Exception):
+    pass