+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## unreleased version -- v0.12.4
 ## Added
 New /fulfill endpoint with ability to check for NYPL login in Bearer authorization header
+Fulfill endpoint returns pre-signed URLs for objects in private buckets when user is logged in
 
 ## unreleased version -- v0.12.4
 ## Added