+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 New /fulfill endpoint with ability to check for NYPL login in Bearer authorization header
 Fulfill endpoint returns pre-signed URLs for objects in private buckets when user is logged in
+JSON from all endpoints (search, OPDS search, etc) includes fulfill URL when object requires NYPL login
 
 ## unreleased version -- v0.12.4
 ## Added